requirements.txt
awslambdaric
boto3
pandas
psycopg2-binary
requests